refactor(web): split file-viewer-panel.tsx into sibling module
Closes #83.

The panel grew to 659 LOC while landing inline markdown/html/image/pdf/
xlsx/office support, violating project CLAUDE.md §5.1's 500-line cap for
new files. Splits the concerns into a sibling directory keyed to
rendering strategy:

  features/messaging/components/file-viewer/
    utils.ts         - Kind type, extension sets, detectKind, typeBadge,
                       parseCsv (pure helpers, already unit-tested)
    csv-table.tsx    - CsvTable primitive
    excel-sheet.tsx  - ExcelSheet + multi-sheet tab bar
    viewer-body.tsx  - per-kind dispatch (image/pdf/office/excel/html/
                       markdown/csv/code/text/unknown)

file-viewer-panel.tsx now holds the header toolbar and three fetch
effects (version list, binary blob, text content) and imports from the
new module. It re-exports detectKind + parseCsv so the existing
file-viewer-panel.test.ts keeps working without an import path change.

File LOC: 659 -> 300 (exactly at cap). No behavioral change, no
className or API surface change. 29/29 messaging tests pass.
