+++ apps/web/features/messaging/components/file-viewer-panel.tsx
@@ -1,66 +1,22 @@
 "use client";
 
 import { useEffect, useMemo, useState } from "react";
-import { ExternalLink, Pencil, Save, X, Eye, Download, RefreshCw } from "lucide-react";
+import { ExternalLink, Pencil, Save, X, Eye, RefreshCw } from "lucide-react";
 import { toast } from "sonner";
 
 import { api } from "@/shared/api";
 import type { FileVersion } from "@/shared/types";
 import { formatSize } from "@/shared/file-display";
-import { MemoizedMarkdown } from "@/components/markdown";
 import type { FileViewerTarget } from "@/features/messaging/stores/file-viewer-store";
+
+import { detectKind, parseCsv, typeBadge } from "./file-viewer/utils";
+import { ViewerBody } from "./file-viewer/viewer-body";
+
+export { detectKind, parseCsv };
 
 interface FileViewerPanelProps {
   target: FileViewerTarget;
   onClose: () => void;
-}
-
-type Kind =
-  | "markdown"
-  | "html"
-  | "text"
-  | "code"
-  | "csv"
-  | "image"
-  | "pdf"
-  | "excel"
-  | "office"
-  | "unknown";
-
-const CODE_EXT = new Set([
-  "ts", "tsx", "js", "jsx", "py", "go", "rs", "java", "c", "cpp", "h",
-  "cs", "rb", "php", "swift", "kt", "scala", "sh", "bash", "zsh",
-  "sql", "toml", "xml", "html", "css", "scss", "vue", "svelte",
-]);
-
-const TEXT_EXT = new Set(["txt", "log", "env", "gitignore"]);
-const DATA_EXT = new Set(["json", "yaml", "yml"]);
-
-export function detectKind(name: string, mime?: string): Kind {
-  const ext = name.split(".").pop()?.toLowerCase() ?? "";
-  if (ext === "md" || ext === "markdown" || mime === "text/markdown") return "markdown";
-  if (ext === "html" || ext === "htm" || mime === "text/html") return "html";
-  if (ext === "csv" || mime === "text/csv") return "csv";
-  if (ext === "pdf" || mime === "application/pdf") return "pdf";
-  if (["png", "jpg", "jpeg", "gif", "webp", "svg", "bmp"].includes(ext)) return "image";
-  if (mime?.startsWith("image/")) return "image";
-  if (ext === "xlsx" || ext === "xls" || mime?.includes("spreadsheetml") || mime === "application/vnd.ms-excel") return "excel";
-  if (["doc", "docx", "ppt", "pptx"].includes(ext)) return "office";
-  if (CODE_EXT.has(ext)) return "code";
-  if (DATA_EXT.has(ext)) return "code";
-  if (TEXT_EXT.has(ext) || mime?.startsWith("text/")) return "text";
-  return "unknown";
-}
-
-function typeBadge(kind: Kind, ext: string) {
-  if (kind === "markdown") return "MD";
-  if (kind === "html") return "HTML";
-  if (kind === "pdf") return "PDF";
-  if (kind === "image") return "IMG";
-  if (kind === "excel") return "XLSX";
-  if (kind === "office") return ext.toUpperCase();
-  if (kind === "csv") return "CSV";
-  return ext.toUpperCase() || "FILE";
 }
 
 export function FileViewerPanel({ target, onClose }: FileViewerPanelProps) {
@@ -341,313 +297,4 @@ export function FileViewerPanel({ target, onClose }: FileViewerPanelProps) {
       </div>
     </div>
   );
-}
-
-interface ViewerBodyProps {
-  kind: Kind;
-  version: FileVersion;
-  content: string;
-  draft: string;
-  setDraft: (v: string) => void;
-  contentLoaded: boolean;
-  contentErr: string;
-  mode: "preview" | "edit";
-  blobUrl: string | null;
-  binaryErr: string;
-  sheets: { name: string; rows: string[][] }[] | null;
-  activeSheetIdx: number;
-  setActiveSheetIdx: (i: number) => void;
-}
-
-function ViewerBody({
-  kind,
-  version,
-  content,
-  draft,
-  setDraft,
-  contentLoaded,
-  contentErr,
-  mode,
-  blobUrl,
-  binaryErr,
-  sheets,
-  activeSheetIdx,
-  setActiveSheetIdx,
-}: ViewerBodyProps) {
-  if (kind === "image") {
-    if (binaryErr) return <div className="p-4 text-[13px] text-destructive">{binaryErr}</div>;
-    if (!blobUrl) return <div className="h-full flex items-center justify-center text-[13px] text-muted-foreground">加载图片中...</div>;
-    return (
-      <div className="p-4 flex items-center justify-center">
-        {/* eslint-disable-next-line @next/next/no-img-element */}
-        <img
-          src={blobUrl}
-          alt={version.filename}
-          className="max-w-full max-h-[85vh] object-contain rounded border border-border"
-        />
-      </div>
-    );
-  }
-
-  if (kind === "pdf") {
-    if (binaryErr) return <div className="p-4 text-[13px] text-destructive">{binaryErr}</div>;
-    if (!blobUrl) return <div className="h-full flex items-center justify-center text-[13px] text-muted-foreground">加载 PDF 中...</div>;
-    return (
-      <iframe
-        src={blobUrl}
-        className="w-full h-full border-0"
-        title={version.filename}
-      />
-    );
-  }
-
-  if (kind === "excel") {
-    if (binaryErr) return <div className="p-4 text-[13px] text-destructive">{binaryErr}</div>;
-    if (!sheets) return <div className="h-full flex items-center justify-center text-[13px] text-muted-foreground">解析表格中...</div>;
-    return (
-      <ExcelSheet sheets={sheets} activeIdx={activeSheetIdx} onSelect={setActiveSheetIdx} />
-    );
-  }
-
-  if (kind === "office") {
-    if (binaryErr) {
-      return (
-        <div className="p-4 text-[13px] text-destructive">
-          {binaryErr}
-        </div>
-      );
-    }
-    return (
-      <div className="h-full flex flex-col">
-        <div className="px-4 py-2 text-[12px] text-muted-foreground bg-secondary/40 border-b border-border flex items-center gap-2">
-          <span>Office 文件暂无内嵌预览，浏览器将尝试下载。</span>
-          {blobUrl && (
-            <a
-              href={blobUrl}
-              download={version.filename}
-              className="inline-flex items-center gap-1 text-primary hover:underline"
-            >
-              <Download className="h-3 w-3" /> 下载
-            </a>
-          )}
-        </div>
-        {blobUrl ? (
-          <iframe src={blobUrl} className="flex-1 w-full border-0" title={version.filename} />
-        ) : (
-          <div className="flex-1 flex items-center justify-center text-[13px] text-muted-foreground">加载中...</div>
-        )}
-      </div>
-    );
-  }
-
-  if (!contentLoaded) {
-    return (
-      <div className="h-full flex items-center justify-center text-[13px] text-muted-foreground">
-        加载内容中...
-      </div>
-    );
-  }
-
-  if (contentErr) {
-    return <div className="p-4 text-[13px] text-destructive">{contentErr}</div>;
-  }
-
-  if (mode === "edit") {
-    return (
-      <textarea
-        value={draft}
-        onChange={(e) => setDraft(e.target.value)}
-        spellCheck={false}
-        className="w-full h-full p-4 bg-background text-foreground font-mono text-[13px] leading-relaxed outline-none resize-none"
-      />
-    );
-  }
-
-  if (kind === "markdown") {
-    return (
-      <div className="p-4 prose prose-sm max-w-none dark:prose-invert">
-        <MemoizedMarkdown mode="full">{content}</MemoizedMarkdown>
-      </div>
-    );
-  }
-
-  if (kind === "html") {
-    // srcDoc so the HTML runs in an isolated document without a network
-    // round-trip. sandbox="allow-scripts" lets inline <script> execute for
-    // real fidelity (diagrams, charts) while blocking access to the parent
-    // app — the iframe is cross-origin and can't read cookies or fire
-    // credentialed requests back to our API.
-    return (
-      <iframe
-        title={version.filename}
-        srcDoc={content}
-        sandbox="allow-scripts"
-        className="w-full h-full border-0 bg-white"
-      />
-    );
-  }
-
-  if (kind === "csv") {
-    return <CsvTable text={content} />;
-  }
-
-  return (
-    <pre className="p-4 text-[13px] leading-relaxed font-mono whitespace-pre-wrap break-words text-foreground">
-      {content}
-    </pre>
-  );
-}
-
-function ExcelSheet({
-  sheets,
-  activeIdx,
-  onSelect,
-}: {
-  sheets: { name: string; rows: string[][] }[];
-  activeIdx: number;
-  onSelect: (i: number) => void;
-}) {
-  if (sheets.length === 0) {
-    return <div className="p-4 text-[13px] text-muted-foreground">工作簿为空</div>;
-  }
-  const sheet = sheets[Math.min(activeIdx, sheets.length - 1)]!;
-  const rows = sheet.rows;
-  const [head, ...body] = rows;
-  const headRow = head ?? [];
-  return (
-    <div className="h-full flex flex-col">
-      {sheets.length > 1 && (
-        <div className="flex items-center gap-1 px-2 py-1 border-b border-border bg-secondary/40 overflow-x-auto shrink-0">
-          {sheets.map((s, i) => (
-            <button
-              key={s.name + i}
-              type="button"
-              onClick={() => onSelect(i)}
-              className={`px-2 py-1 text-[12px] rounded-[4px] transition-colors whitespace-nowrap ${
-                i === activeIdx
-                  ? "bg-background text-foreground shadow-sm"
-                  : "text-muted-foreground hover:bg-accent/50 hover:text-foreground"
-              }`}
-            >
-              {s.name}
-            </button>
-          ))}
-        </div>
-      )}
-      <div className="flex-1 overflow-auto p-4">
-        {rows.length === 0 ? (
-          <div className="text-[13px] text-muted-foreground">该工作表为空</div>
-        ) : (
-          <table className="min-w-full text-[12px] border-collapse">
-            <thead>
-              <tr>
-                {headRow.map((cell, i) => (
-                  <th
-                    key={i}
-                    className="border border-border bg-secondary px-2 py-1 text-left font-medium text-foreground sticky top-0"
-                  >
-                    {cell}
-                  </th>
-                ))}
-              </tr>
-            </thead>
-            <tbody>
-              {body.map((row, r) => (
-                <tr key={r} className="hover:bg-accent/30">
-                  {row.map((cell, c) => (
-                    <td key={c} className="border border-border px-2 py-1 text-foreground align-top">
-                      {cell}
-                    </td>
-                  ))}
-                </tr>
-              ))}
-            </tbody>
-          </table>
-        )}
-      </div>
-    </div>
-  );
-}
-
-function CsvTable({ text }: { text: string }) {
-  const rows = useMemo(() => parseCsv(text), [text]);
-  if (rows.length === 0) {
-    return <div className="p-4 text-[13px] text-muted-foreground">空文件</div>;
-  }
-  const [head, ...body] = rows;
-  const headRow = head ?? [];
-  return (
-    <div className="p-4 overflow-auto">
-      <table className="min-w-full text-[12px] border-collapse">
-        <thead>
-          <tr>
-            {headRow.map((cell, i) => (
-              <th
-                key={i}
-                className="border border-border bg-secondary px-2 py-1 text-left font-medium text-foreground"
-              >
-                {cell}
-              </th>
-            ))}
-          </tr>
-        </thead>
-        <tbody>
-          {body.map((row, r) => (
-            <tr key={r} className="hover:bg-accent/30">
-              {row.map((cell, c) => (
-                <td key={c} className="border border-border px-2 py-1 text-foreground align-top">
-                  {cell}
-                </td>
-              ))}
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
-  );
-}
-
-// Small CSV parser — handles quoted fields with embedded commas/newlines and
-// escaped quotes ("" → "). Good enough for preview; not a full RFC 4180
-// implementation.
-export function parseCsv(text: string): string[][] {
-  const rows: string[][] = [];
-  let row: string[] = [];
-  let field = "";
-  let inQuotes = false;
-  for (let i = 0; i < text.length; i++) {
-    const c = text[i];
-    if (inQuotes) {
-      if (c === '"') {
-        if (text[i + 1] === '"') {
-          field += '"';
-          i++;
-        } else {
-          inQuotes = false;
-        }
-      } else {
-        field += c;
-      }
-      continue;
-    }
-    if (c === '"') {
-      inQuotes = true;
-    } else if (c === ",") {
-      row.push(field);
-      field = "";
-    } else if (c === "\n" || c === "\r") {
-      if (c === "\r" && text[i + 1] === "\n") i++;
-      row.push(field);
-      field = "";
-      rows.push(row);
-      row = [];
-    } else {
-      field += c;
-    }
-  }
-  if (field.length > 0 || row.length > 0) {
-    row.push(field);
-    rows.push(row);
-  }
-  return rows.filter((r) => r.some((c) => c.length > 0));
 }
